backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
numpy
mediapipe
opencv-python-headless
